chore: Move global "TrackToPlaylistSheet" to its own file
